PostManager Code Edited

+++ src/main/java/ru/netology/manager/PostManager.java
@@ -5,7 +5,7 @@ import ru.netology.domain.Post;
 public class PostManager {
     private Post[] posts;
 
-    public Post[] search(int ownerId, String domain, String query, boolean ownersOnly, int count, long offset) {
+    public Post[] search(int ownerId, String domain, String query, boolean ownersOnly, int count, int offset) {
         // logic
         return null;
     }
